+++ src/hmm/hmm_estimateB.c
@@ -3,12 +3,12 @@
 
 #include "hmm.h"
 #include "utl.h"
+#include "distr.h" // dis_inicUni
 
 #include <stdlib.h>
 
-static const prob_t one = (prob_t) 1.;
 
-void hmm_estimateB(Hmm *hmm, Symbol **seqs, int *T, int num_cads, int max_T) {
+void hmm_estimateB(Hmm *hmm, Symbol **seqs, int *Ts, int num_seqs, int max_T) {
     printf("estimating initial B matrix ...  (given max_T=%d)\n", max_T);
     const double measure_start_sec = measure_time_now_sec();
 
@@ -18,8 +18,9 @@ void hmm_estimateB(Hmm *hmm, Symbol **seqs, int *T, int num_cads, int max_T) {
 
     if (max_T <= 0) {
         max_T = 0;
-        for (int r = 0; r < num_cads; ++r) {
-            if (max_T < T[r]) max_T = T[r];
+        for (int r = 0; r < num_seqs; ++r) {
+            const int T = Ts[r];
+            if (max_T < T) max_T = T;
         }
         printf("(computed max_T=%d)\n", max_T);
     }
@@ -42,18 +43,21 @@ void hmm_estimateB(Hmm *hmm, Symbol **seqs, int *T, int num_cads, int max_T) {
     prob_t **phi = (prob_t **) new_matrix(max_T, N, sizeof(prob_t));
     int **psi = (int **) new_matrix(max_T, N, sizeof(int));
 
-    for (int r = 0; r < num_cads; r++) {
-        hmm_precompute_logB(hmm, seqs[r], T[r], logB);
+    for (int r = 0; r < num_seqs; r++) {
+        Symbol *O = seqs[r];
+        const int T = Ts[r];
+
+        hmm_precompute_logB(hmm, O, T, logB);
 
         // get optimal sequence:
-        hmm_genQopt_with_mem(hmm, seqs[r], T[r], Qopt,
+        hmm_genQopt_with_mem(hmm, O, T, Qopt,
                 logA, logB,
                 phi, psi);
 
         // update counters:
-        for (int t = 0; t < T[r]; t++) {
+        for (int t = 0; t < T; t++) {
             const int state = Qopt[t];
-            num[state][seqs[r][t]]++;
+            num[state][O[t]]++;
             den[state]++;
         }
     }
@@ -64,14 +68,17 @@ void hmm_estimateB(Hmm *hmm, Symbol **seqs, int *T, int num_cads, int max_T) {
 
     del_vector(Qopt);
 
+    // TODO apply epsilon restriction only in relevant rows below
+
+    int num_not_emitting_states = 0;
+
     // do required normalizations:
     for (int j = 0; j < N; j++) {
         // no emitted symbol from this state?
         if (den[j] == 0) {
             // set uniform distribution:
-            for (int k = 0; k < M; k++) {
-                B[j][k] = one / M;
-            }
+            dis_inicUni(B[j], M);
+            ++num_not_emitting_states;
         }
         else {
             // set B[j] with normalized distribution:
@@ -84,8 +91,9 @@ void hmm_estimateB(Hmm *hmm, Symbol **seqs, int *T, int num_cads, int max_T) {
     del_vector(den);
     del_matrix(num);
 
+    printf("num_not_emitting_states=%d\n", num_not_emitting_states);
     printf("initial B matrix took %s\n", measure_time_show_elapsed(measure_time_now_sec() - measure_start_sec));
 
-    // now apply epsilon-restriction:
+    // now apply epsilon-restriction:  TODO see above
     hmm_adjustB(hmm, "hmm_estimateB");
 }